backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-whois
dnspython
